[REFACTOR] remove mixins and make them into main and templates
